Move Command implementations to impl package
