+++ src/main/java/com/elsria/commands/CommandType.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import com.elsria.commands.impl.*;
 import com.elsria.core.ApplicationContext;
 
 /**
